requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
geoalchemy2==0.14.0
pgvector==0.2.0
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-settings==2.0.3
sentence-transformers==2.2.2
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
httpx==0.24.1
crewai==0.28.3
google-api-python-client==2.108.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
cachetools==5.3.2
redis==5.0.1
tenacity==8.2.3 